+++ actions/dashboard/users/usersActions.ts
@@ -1,10 +1,21 @@
 "use server";
 
 import prisma from "@/lib/db";
+import { getCurrentUser } from "@/lib/getCurrentUser";
 import { Role } from "@prisma/client";
 import { revalidatePath } from "next/cache";
 
 export async function deleteUser(id: string) {
+
+  const currentUser = await getCurrentUser();
+
+  if (currentUser?.role !== "SUPER_ADMIN") {
+    return {
+      status: 400,
+      message: "You don't have permission to delete this user.",
+    };
+  }
+
   try {
     await prisma.user.delete({
         where:{id},
@@ -28,6 +39,16 @@ export async function deleteUser(id: string) {
 
 
 export async function changeUserStatus(userId:string, role:Role){
+
+  const currentUser = await getCurrentUser();
+
+  if(currentUser?.role !== "SUPER_ADMIN"){
+    return {
+      status: 400,
+      message: "You don't have permission to change user role.",
+    };
+  }
+
   try {
     await prisma.user.update({
       where:{
